stack sucks
stack install github-backup fails when run outside the source tree. For
some reason, stack does not look at the stack.yaml when used that way,
and is left trying to use whatever the default lts version is, which
likely fails. So, adjust documentation to avoid users expericing this
broken stack behavior.