Updated to use github-0.15.
Old versions not supported, although it would not be too hard to ifdef
support back in.

Github.issuesForRepo' API has improved; it can be called once now with
stateAll to get get both open and closed issues.

However, pullRequestsFor' is gone; temporarily switch to using
pullRequestsFor, which does not allow an Auth to be provided.
https://github.com/phadej/github/issues/266

Lightly tested.

This commit was sponsored by Jeff Goeke-Smith on Patreon.